MiniProject/hospital-management-app/src/main/java/com/prachi/app/controller/DoctorController.java
package com.prachi.app.controller;


import com.prachi.app.model.Appointment;
import com.prachi.app.repository.AppointmentRepository;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.web.bind.annotation.*;

import java.util.List;


@RestController
@RequestMapping("/doctor")
public class DoctorController {

    @Autowired
    AppointmentRepository appoinmentRepo;


    @GetMapping("/doctorappointment")
    public List<Appointment> getAppointments(@RequestParam String doctorName){
        return appoinmentRepo.findByDoctorName(doctorName);
    }

    @PostMapping("/save")
    public Appointment saveAppointment(@RequestBody Appointment appointment){
        appointment = appoinmentRepo.save(appointment);
        return appointment;
    }

}
